ZKL ChatControl(s) :allow battleroom minimap to be resizeable (add another splitContainer). BattleChatControl.cs: expand that new container when in battleroom mode^^
Utils.cs: dispose Graphic for measuring dpi that is used only once.
NotifyBarContainer.cs: insignificant tweak, double to int.

svn path=/trunk/other/; revision=9655

+++ ZeroKLobby/Notifications/NotifyBarContainer.cs
@@ -19,9 +19,9 @@ namespace ZeroKLobby.Notifications
             
             //-- code for countering the 'DPI scaling cause bar to grow infinitely big'. The bar grew larger because each time the bar was created the bar is scaled to high-DPI, but since old height is used again for new bar: the bar get bigger & bigger after each scaling.
             DpiMeasurement.DpiXYMeasurement(this); //this measurement use cached value. It won't cost anything if another measurement was already done in other control element
-            double newHeight = DpiMeasurement.ReverseScaleValueY(control.Height); //get old height. Note: DpiMeasurement is a static class stored in ZeroKLobby\Util.cs
+            int newHeight = DpiMeasurement.ReverseScaleValueY(control.Height); //get old height. Note: DpiMeasurement is a static class stored in ZeroKLobby\Util.cs
             //--
-            Height = (int) newHeight + 8; // apply height + margin (the height will again be automatically be scaled up according to Window's DPI) 
+            Height = newHeight + 8; // apply height + margin (the height will again be automatically be scaled up according to Window's DPI) 
             tableLayoutPanel1.Controls.Add(control, 1, 0);
             control.Dock = DockStyle.Fill;
             Dock = DockStyle.Top;
